04-RAG/01-RAG-Application-With-FAISS/requirements.txt
streamlit
faiss-cpu
numpy
requests
python-dotenv